Finish coding SmppServer and SmppSession classes. Add some tests.
SmppServer: was missing the initialization of the status variable and to close the server socket at the end.
SmppSession: create a constructor that receives a StreamLink and uses a default PacketProcessor.
SmppSession: create a new sendRequest method to send requests to the client.
SmppSession: second bind request should fail.
SmppSession: request without a ind should fail.
Create tests for SmppSession and finish the tests of SmppServer.